\name{qNorm}
\alias{qNorm}
\title{Quantile transform counts}
\usage{
  qNorm(x, verbose = FALSE)
}
\arguments{
  \item{x}{raw counts}

  \item{verbose}{print head of transformed matrix if TRUE
  (default=FALSE)}
}
\value{
  matrix of quantile normalized counts
}
\description{
  Quantile transform counts
}

