% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{qNorm}
\alias{qNorm}
\title{Quantile transform counts.}
\usage{
qNorm(x, verbose = FALSE)
}
\arguments{
\item{x}{raw counts}

\item{verbose}{print head of transformed matrix if TRUE (default=FALSE)}
}
\value{
matrix of quantile normalized counts
}
\description{
Perform normalize.quantiles on the count matrix, likely used with log2CPM().
}

