% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{voomMod}
\alias{voomMod}
\title{Call modified voom on ComBatLoc adjusted data.}
\usage{
voomMod(x, design, lib.size)
}
\arguments{
\item{x}{ComBatLoc adjusted data}

\item{design}{reparameterized design matrix}

\item{lib.size}{library sizes of quantile counts obtained from calling logPCM()}
}
\value{
Elist object including a plot slot containing a ggplot2 version of voom()'s plot.
}
\description{
This is a copy of limma's voom() with a couple changes to allow the input of log2(cpm()) data.
}

